Split js web app into library
